Revert "Release v5.0.0-rc6 (#4736)" (#4740)
This reverts commit 6ceb82d8fb0f577daaecad83d34c8ce391c8c7dd.



<!--
## Before submiting a pull request

Please make sure you've run the following commands from the root
directory.

    bin/generate-all.sh

(this runs commands like "go generate", fixes formatting, and so on)

## Release changelog section

Help keep the release changelog clear by pre-naming the proper section
in the GitHub pull request title.

Some examples:
* CICD: Add required GHA permissions for goreleaser
* DOCS: Fixed providers with "contributor support" table
* ROUTE53: Allow R53_ALIAS records to enable target health evaluation

More examples/context can be found in the file .goreleaser.yml under the
'build' > 'changelog' key.
!-->

<!-- codesmith:footer -->
---
<a
href="https://app.blacksmith.sh/DNSControl/codesmith/dnscontrol/pr/4740"><picture><source
media="(prefers-color-scheme: dark)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-dark-v2.svg"><source
media="(prefers-color-scheme: light)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-light-v2.svg"><img
alt="View with [code]smith"
src="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-dark-v2.svg"></picture></a>
<a
href="https://backend.blacksmith.sh/track/enable-autofix?expires=1789071775&installation_model_id=431813&pr_number=4740&repository=DNSControl%2Fdnscontrol&return_to=https%3A%2F%2Fgithub.com%2FDNSControl%2Fdnscontrol%2Fpull%2F4740&signature=b594942e5d1726ddbf0c6b8e5a6f4bb992ae5045e58c962d91c3e8b179945f94"><picture><source
media="(prefers-color-scheme: dark)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-dark.svg"><source
media="(prefers-color-scheme: light)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-light.svg"><img
alt="Autofix with [code]smith"
src="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-dark.svg"></picture></a>
<sup>Need help on this PR? Tag <code>@codesmith-bot</code> with what you
need. Autofix is disabled.</sup>

<!-- codesmith:autofix:disabled -->
<!-- /codesmith:footer -->

+++ providers/bunnydns/records.go
@@ -1,12 +1,13 @@
 package bunnydns
 
 import (
+	"errors"
 	"fmt"
 	"slices"
 
-	"github.com/DNSControl/dnscontrol/v5/models"
-	"github.com/DNSControl/dnscontrol/v5/pkg/diff2"
-	"github.com/DNSControl/dnscontrol/v5/pkg/printer"
+	"github.com/DNSControl/dnscontrol/v4/models"
+	"github.com/DNSControl/dnscontrol/v4/pkg/diff2"
+	"github.com/DNSControl/dnscontrol/v4/pkg/printer"
 )
 
 func (b *bunnydnsProvider) GetZoneRecords(dc *models.DomainConfig) (models.Records, error) {
@@ -22,7 +23,13 @@ func (b *bunnydnsProvider) GetZoneRecords(dc *models.DomainConfig) (models.Recor
 		return nil, err
 	}
 
-	recs := make(models.Records, 0, len(nativeRecs))
+	implicitRecs, err := b.getImplicitRecordConfigs(zone)
+	if err != nil {
+		return nil, err
+	}
+
+	recs := make(models.Records, 0, len(nativeRecs)+len(implicitRecs))
+	recs = append(recs, implicitRecs...)
 
 	// Define a list of record types that are currently not supported by this provider.
 	unsupportedTypes := []recordType{
@@ -38,7 +45,7 @@ func (b *bunnydnsProvider) GetZoneRecords(dc *models.DomainConfig) (models.Recor
 			continue
 		}
 
-		rc, err := toRecordConfig(dc, nativeRec)
+		rc, err := toRecordConfig(zone.Domain, nativeRec)
 		if err != nil {
 			return nil, err
 		}
@@ -126,7 +133,11 @@ func (b *bunnydnsProvider) mkChangeCorrection(zoneID int64, oldRec, newRec *mode
 	return &models.Correction{
 		Msg: msg,
 		F: func() error {
-			existingID := oldRec.Original.(int64)
+			existingID := oldRec.Original.(*record).ID
+			if existingID == 0 {
+				return errors.New("BUNNY_DNS: cannot change implicit records")
+			}
+
 			desired, err := fromRecordConfig(newRec)
 			if err != nil {
 				return err
@@ -141,7 +152,11 @@ func (b *bunnydnsProvider) mkDeleteCorrection(zoneID int64, oldRec *models.Recor
 	return &models.Correction{
 		Msg: msg,
 		F: func() error {
-			existingID := oldRec.Original.(int64)
+			existingID := oldRec.Original.(*record).ID
+			if existingID == 0 {
+				return errors.New("BUNNY_DNS: cannot delete implicit records")
+			}
+
 			return b.deleteRecord(zoneID, existingID)
 		},
 	}
